Change the cursor type to have an explicit list
We decided that the older version of the cursor type probably would not
work. As a result, we added a type to Core that represents a list of
types, and we use this type for the Needs and Has cursors now. The
definition of `typeOfPrimFunArity` and the functions that depend on it
are now in Helium as well, since the `PrimWriteCtor` function requires
the type environment to work.

As for why all of the Type.hs file has been changed: I had some trouble
with line-endings and the result is that I changed all line-endings to
Unix style. Usually Git ends up fixing it, but not this time I guess.