Add WriteLength primitive and Pattern Synonyms
The WriteLength primitive is for the cursors. Cursors now need to keep
track of when to write a length of a new field. This type tag is
supposed to help with that.

The pattern synonyms are more generally useful. There is one for lists
of types, similar to what (:) does for values in normal Haskell, which
operates on the TConCons. The other is for function types, and operates
on the TConFun constructor.